project: Use CMakeLists.tmpl from Mbed OS
The project CMakeLists.txt template has changed in different versions
of Mbed OS. So that all CLI2 supported versions of Mbed OS work with
all future versions of the tools, this should be decoupled, with new
versions of Mbed OS including the required template.

This generates the CMakeLists.txt file after initialising with new,
using the template from Mbed OS if available. If no template is
available, the tools will fall back on a local copy.